#187: Move functions to helpers.ts
